Add debug-mode install script and fix bugs

+++ locks_test/bench.h
@@ -3,7 +3,7 @@
 extern struct bench *bench;
 
 struct worker {
-	unsigned int works;
+	unsigned long long int works;
 };
 
 struct bench {
